feat(incr7): add error type for handlers - rename files handlers - add middleware json - add message for errors - rename methods by handlers - add package respond for handlers - edit go.mod
