+++ internal/http/middleware/json.go
@@ -1,0 +1,10 @@
+package middleware
+
+import "net/http"
+
+func Json(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Content-Type", "application/json")
+		next.ServeHTTP(w, r)
+	})
+}
