+++ internal/http/routes.go
@@ -8,14 +8,18 @@ import (
 // Routes returns the initialized router
 func (s *Server) useRoutes() *chi.Mux {
 	r := chi.NewRouter()
-	r.Use(middleware.Logger)
 
-	r.Post("/", s.handler.PostWithTextData)
-	r.Get("/{id}", s.handler.GetWithTextData)
+	r.Use(
+		middleware.Json,
+		middleware.Logger,
+	)
+
+	r.Post("/", s.handler.Post)
+	r.Get("/{id}", s.handler.Get)
 
 	r.Route("/api", func(r chi.Router) {
 		r.Route("/shorten", func(r chi.Router) {
-			r.Post("/", s.handler.PostWithDataJSON)
+			r.Post("/", s.handler.PostAPI)
 		})
 	})
